+++ src/helpers/hooks/useSelector.ts
@@ -1,6 +1,6 @@
 import { useSelector, shallowEqual } from 'react-redux';
 
-import { RootState } from '@/store';
+import { RootState } from '@/store/rootState';
 
 type TStateSelector<T> = (state: RootState) => T;
 type TShallowEq = typeof shallowEqual;
