+++ src/screens/Article/Article.tsx
@@ -7,9 +7,9 @@ import { RouteProp, useRoute } from '@react-navigation/native';
 import { initialArticle, initialChatInfo } from '@/constants/InitialState';
 import { createError } from '@/helpers/functions';
 import AppLoading from '@/screens/AppLoading/AppLoading';
-import { RootState } from '@/store';
-import { getSingleArticle } from '@/store/articleSlice';
-import { getChatInfo } from '@/store/chatSlice';
+import { getSingleArticle } from '@/store/articleSlice/articleSlice';
+import { getChatInfo } from '@/store/chatSlice/chatSlice';
+import { RootState } from '@/store/rootState';
 import { IArticleProps } from '@/types/article';
 import { IChattingRoom } from '@/types/chat';
 import { ArticleDrawerParamList } from '@/types/navigation';
