+++ src/screens/Drawer/Drawer.tsx
@@ -12,8 +12,8 @@ import { Button, Profile } from '@/components';
 import { Status } from '@/constants/Enum';
 import { asyncStoragekey } from '@/constants/asyncStorage';
 import { ObjectStorage } from '@/helpers/functions/asyncStorage';
-import { RootState } from '@/store';
-import { getChatInfo, changeOrderStatus } from '@/store/chatSlice';
+import { getChatInfo, changeOrderStatus } from '@/store/chatSlice/chatSlice';
+import { RootState } from '@/store/rootState';
 import { palette } from '@/styles';
 import { IChattingRoom } from '@/types/chat';
 import { IUserProps } from '@/types/user';
