+++ src/screens/Profile/Profile.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
 
-import { RootState } from '@/store';
+import { RootState } from '@/store/rootState';
 
 import LoggedProfile from './Logged/Logged';
 import UnloggedProfile from './UnLogged/UnLogged';
