+++ src/store/rootStore.ts
@@ -1,6 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit';
 
-import rootReducer from '.';
+import rootReducer from './rootState';
 
 export type AppDispatch = typeof store.dispatch;
